indent(1): Fix buildwork
Fix
error: comparison of integers of different signs: 'int' and 'unsigned long'
by casting nitems() to int.

+++ usr.bin/indent/parse.c
@@ -241,7 +241,7 @@ parse(int tk)
 
 	}			/* end of switch */
 
-	if (ps.tos >= nitems(ps.p_stack) - 1)
+	if (ps.tos >= (int)nitems(ps.p_stack) - 1)
 		errx(1, "Parser stack overflow.")
 	
 	reduce();		/* see if any reduction can be done */
